+++ packages/cli/src/commands/generate/service/streaming.ts
@@ -1,0 +1,7 @@
+import { Client } from "../../../util/client.js";
+
+export const generateStreamingService = async (client: Client) => {
+  const { output } = client;
+
+  console.log("generating streaming service");
+};