+++ packages/create-wave-app/CHANGELOG.md
@@ -1,5 +1,35 @@
 # create-wave-app
 
+## 0.10.15
+
+### Patch Changes
+
+- Create frontend src folder for template, update wave-cli version
+
+## 0.10.14
+
+### Patch Changes
+
+- Add graphql codegen plugin, update wave cli version for create-wave-app templates
+
+## 0.10.13
+
+### Patch Changes
+
+- added devDeps to the correct package.json files in create-wave-app
+
+## 0.10.12
+
+### Patch Changes
+
+- removes devDeps adds deps to create-wave-app package.json
+
+## 0.10.11
+
+### Patch Changes
+
+- added deps and devDeps to root package.json
+
 ## 0.10.10
 
 ### Patch Changes